Initial pass at time-to-plane
MechJeb has a nice feature which allows the user to time their launch
to occur in the same orbital plane as a target.  This makes rendezvous
missions cost less fuel and is pretty keen.

GT already has support for using MJ functionality, so I added support
for MJ's time-to-plane method.

I added a check for MechJeb and a target being set to the stats window
to only display the time-to-plane and inclination when appropriate.  The
stats window output is ugly but temporary.

To test the feature:
1. Load up a new sandbox world with MJ installed.
2. Pick "Orbital One" in the VAB and slap a MJ part on it.
3. Go to the launchpad with it and open GT.  Enable the stats window.
Note there is no reference to the target, because it isn't yet set.
4. Go to the map view and set Minmus as your target.
Note that the window now shows a time-to-plane of about an hour, and
an inclination of 6.00 degrees.
5. Exit the map view, change the GT inclination to 6.00 degrees, and
warp time until the time-to-plane is less than 15 seconds.
6. Count down until time-to-plane is zero, and hit 'launch'.

When your circularization burn is complete, use MJ to align planes with
the target.  Your burn should be on the order of 5-15 m/s.